% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decon_single_sample.R
\name{Decon_single_sample}
\alias{Decon_single_sample}
\title{Function for single sample weights estimation}
\usage{
Decon_single_sample(refSet, data, geneIDType)
}
\arguments{
\item{refSet}{Label of pre-stored deconvoluted reference dataset}

\item{data}{Data matrix to be deconvoluted with gene ID as rownames}

\item{geneIDType}{Choose one among "geneSymbol", "EntrezID" and "TCGA"}
}
\value{
A matrix of the estimated sample compartment weights
}
\description{
This function allows you to calculate singel sampmle weights
}
