% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_PDAC_weights.R
\name{Norm_PDAC_weights}
\alias{Norm_PDAC_weights}
\title{Function for sample weight normalization for PDAC}
\usage{
Norm_PDAC_weights(sampleWeight)
}
\arguments{
\item{sampleWeight}{Sample weights calculated by Decon_single_sample.R}
}
\value{
Normalized sample weights
}
\description{
Function for sample weight normalization for PDAC
}
